pyproject.toml

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cifix"
version = "0.4.0"
description = "A CLI tool built with Click"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]
anthropic = [
    "anthropic>=0.39.0",
]
openai = [
    "openai>=1.0",
]
gemini = [
    "google-generativeai>=0.8.0",
]
llm = [
    "anthropic>=0.39.0",
    "openai>=1.0",
    "google-generativeai>=0.8.0",
]

[project.scripts]
cifix = "cifix.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
